Fix issues with enableif components
Sync with commit 74d8907 of NavtableForms to fix the issues with
enableif components when filling values. 'cod_bombeo' in
'captaciones' and 'dep_intermedios' were generating problems. Also,
the foreign key of these fields on the 'bombeos' table has been
removed, because when the slave cb is disabled its value becomes a
blank string, and not a null.