Added to the factory a method for loading layers.
Now it has a method that checks whether a layers is
loaded, and if it isn't, it tries to load it with
the current db session.
Maybe the factory was not the better place for
this function, but I don't think we have a better
place right now without creating a new class.