VectorialTable working with models w/o foreignkey
Now the VectorialTableHandler can filter tables
using columns that are not explicitly set in the
model, as long as the field name exists in the layer.
It adds the column, filters by it and then removes it.
If the foreign key column is in the model, then the
column stays after filtering.